Refactor import paths for components.
